Shortening gifs links and changing store modules tests files names
